+++ src/Application/Query/Happening/Webinar/GetWebinarViewQueryHandler.php
@@ -106,7 +106,8 @@ class GetWebinarViewQueryHandler
             $this->dateTime,
             $timeRemainingInSeconds,
             round($timeRemainingInSeconds * 0.2),
-            $happening->getWebinarHeaderImage($query->getLocale())
+            $happening->getWebinarHeaderImage($query->getLocale()),
+            $happening->getLiveUrl()
         );
     }
 }