+++ tests/Ui/Bundle/AdminBundle/Controller/Happening/UpdateActionTest.php
@@ -162,6 +162,7 @@ class UpdateActionTest extends TestCase
         $this->happening->getInvitationCode()->shouldBeCalled()->willReturn('toto');
         $this->happening->isWebinar()->shouldBeCalled()->willReturn(false);
         $this->happening->isInteractiveWebinar()->shouldBeCalled()->willReturn(false);
+        $this->happening->getLiveUrl()->shouldBeCalled()->willReturn(null);
 
         $form = $this->prophesize(Form::class);
         $formView = $this->prophesize(FormView::class);
@@ -234,6 +235,7 @@ class UpdateActionTest extends TestCase
         $this->happening->getInvitationCode()->shouldBeCalled()->willReturn('toto');
         $this->happening->isWebinar()->shouldBeCalled()->willReturn(false);
         $this->happening->isInteractiveWebinar()->shouldBeCalled()->willReturn(false);
+        $this->happening->getLiveUrl()->shouldBeCalled()->willReturn(null);
 
         $form = $this->prophesize(Form::class);
         $update = new Update($this->happening->reveal());